Ember Particles
I have created a system in which embers from the torch will go into the direction of the campfire. Right now they go the entire path but plan to adjust it to where they fade out before reaching the fire. All changes are in "Matthew's Scene" and can be adjusted as needed.